chore: change the logic for finding configuration files
