+++ configstack/module.go
@@ -11,6 +11,7 @@ import (
 
 	"github.com/gruntwork-io/go-commons/collections"
 	"github.com/gruntwork-io/go-commons/errors"
+	"github.com/gruntwork-io/go-commons/files"
 	"github.com/gruntwork-io/terragrunt/config"
 	"github.com/gruntwork-io/terragrunt/options"
 	"github.com/gruntwork-io/terragrunt/shell"
@@ -480,6 +481,10 @@ func getDependenciesForModule(module *TerraformModule, moduleMap map[string]*Ter
 			return dependencies, nil
 		}
 
+		if files.FileExists(dependencyModulePath) && !files.IsDir(dependencyModulePath) {
+			dependencyModulePath = filepath.Dir(dependencyModulePath)
+		}
+
 		dependencyModule, foundModule := moduleMap[dependencyModulePath]
 		if !foundModule {
 			err := UnrecognizedDependency{
@@ -571,37 +576,6 @@ func FindWhereWorkingDirIsIncluded(terragruntOptions *options.TerragruntOptions,
 	}
 
 	return matchedModules
-}
-
-// Custom error types
-
-type UnrecognizedDependency struct {
-	ModulePath            string
-	DependencyPath        string
-	TerragruntConfigPaths []string
-}
-
-func (err UnrecognizedDependency) Error() string {
-	return fmt.Sprintf("Module %s specifies %s as a dependency, but that dependency was not one of the ones found while scanning subfolders: %v", err.ModulePath, err.DependencyPath, err.TerragruntConfigPaths)
-}
-
-type ErrorProcessingModule struct {
-	UnderlyingError       error
-	ModulePath            string
-	HowThisModuleWasFound string
-}
-
-func (err ErrorProcessingModule) Error() string {
-	return fmt.Sprintf("Error processing module at '%s'. How this module was found: %s. Underlying error: %v", err.ModulePath, err.HowThisModuleWasFound, err.UnderlyingError)
-}
-
-type InfiniteRecursion struct {
-	RecursionLevel int
-	Modules        map[string]*TerraformModule
-}
-
-func (err InfiniteRecursion) Error() string {
-	return fmt.Sprintf("Hit what seems to be an infinite recursion after going %d levels deep. Please check for a circular dependency! Modules involved: %v", err.RecursionLevel, err.Modules)
 }
 
 // ForceLogLevelHook - log hook which can change log level for messages which contains specific substrings
